+++ packages/react-native-sdk/CHANGELOG.md
@@ -2,6 +2,19 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [1.9.1](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-1.9.0...@stream-io/video-react-native-sdk-1.9.1) (2025-01-16)
+
+### Dependency Updates
+
+* `@stream-io/video-client` updated to version `1.15.1`
+* `@stream-io/video-react-bindings` updated to version `1.4.1`
+## [1.9.0](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-1.8.0...@stream-io/video-react-native-sdk-1.9.0) (2025-01-16)
+
+
+### Features
+
+* android 12+ pip uses setAutoEnterEnabled api ([#1643](https://github.com/GetStream/stream-video-js/issues/1643)) ([b07a9a6](https://github.com/GetStream/stream-video-js/commit/b07a9a6a2d97fded37161cfbabc4d9a73baae26a))
+
 ## [1.8.0](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-1.7.0...@stream-io/video-react-native-sdk-1.8.0) (2025-01-15)
 
 ### Dependency Updates